content/canvas/test/test_2d.gradient.object.invalidcolour.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.gradient.object.invalidcolour</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var g = ctx.createLinearGradient(0, 0, 100, 0);
var _thrown = undefined; try {
  g.addColorStop(0, "");
} catch (e) { _thrown = e }; ok(_thrown && _thrown.code == DOMException.SYNTAX_ERR, "should throw SYNTAX_ERR");
var _thrown = undefined; try {
  g.addColorStop(0, 'undefined');
} catch (e) { _thrown = e }; ok(_thrown && _thrown.code == DOMException.SYNTAX_ERR, "should throw SYNTAX_ERR");

if (!_deferred) SimpleTest.finish();
});
</script>
